add author page
Rouge draft complete. I have a Posts page which displays the title and
body of all posts, a Post page which displays the author as a link and
a list of comments in addition to the title and body of the post, and
an Author page which displays only the posts for the current author.